# backend/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is a sample server api for beacon based attendance system"
  version: "1.0.0"
  title: "Swagger beacon-based attendance system"
  contact:
    email: "mdsazzadul.islam@tuni.fi"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "127.0.0.1:8080"
basePath: "/v1"
paths:
...
  /beaconRoom/{roomId}:
    post:
      tags:
      - "association"
      summary: "add association between room and beacon"
      description: "Add a new association between Room and Beacon"
      operationId: "addRoomBeaconAssociation"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      - in: "path"
        name: "roomId"
        required: true
        type: "integer"
        minimum: 1
        format: "int32"
        description: "ID of the room"
      - in: "body"
        name: "body"
        description: "beacons ids that needs to be added to the database"
        required: true
        schema:
          $ref: "#/definitions/beaconIds"
      responses:
        201:
          description: "successful operation"
          schema:
            $ref: '#/definitions/responseWithMessage'
        400:
          description: "Invalid ID supplied"
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      x-swagger-router-controller: "Association"
    get:
      tags:
      - "association"
      summary: "get beacons in a room"
      description: "Get all beacons associated in a room"
      operationId: "getBeaconInRoom"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      - in: "path"
        name: "roomId"
        description: "ID of room"
        required: true
        type: "integer"
        minimum: 1
        format: "int32"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: '#/definitions/beaconsList'
        400:
          description: "Invalid ID supplied"
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      x-swagger-router-controller: "Association"
  /beaconRoom:
    get:
      tags:
      - "association"
      summary: "get all associations"
      description: "get all beacon and room associations"
      operationId: "getAllAssociations"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: '#/definitions/allAssociations'
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      x-swagger-router-controller: "Association"
    post:
      tags:
      - "association"
      summary: "add association between room and beacon"
      description: "Add a new association between Room and Beacon"
      operationId: "addAssociation"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "one room id and one beacon id"
        required: true
        schema:
          $ref: "#/definitions/roomIdBeaconId"
      responses:
        201:
          description: "successful operation"
          schema:
            $ref: '#/definitions/responseWithMessage'
        400:
          description: "Invalid ID supplied"
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      x-swagger-router-controller: "Association"
    delete:
      tags:
      - "association"
      summary: "delete an association"
      description: "Delete an existing association between room and beacon"
      operationId: "deleteAssociation"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "one room id and one beacon id"
        required: true
        schema:
          $ref: "#/definitions/roomIdBeaconId"
      responses:
        201:
          description: "successful operation"
          schema:
            $ref: '#/definitions/responseWithMessage'
        400:
          description: "Invalid ID supplied"
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      x-swagger-router-controller: "Association"
...
  /attendance/searchBeacon:
    post:
      tags:
      - "attendance"
      summary: "Filter from a list of beacons"
      operationId: "searchBeacon"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Course object in body property"
        required: true
        schema:
          $ref: '#/definitions/arrayBeaconsApp'
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: '#/definitions/filteredBeacons'
        400:
          description: "Invalid data"
          schema:
            $ref: '#/definitions/responseWithMessage'
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      x-swagger-router-controller: "Attendance"
  /attendance/takeAttendance:
    post:
      tags:
      - "attendance"
      summary: "Take attendance from student"
      operationId: "takeAttendance"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Course object in body property"
        required: true
        schema:
          $ref: '#/definitions/takeAttendanceBody'
      responses:
        201:
          description: "successful operation"
          schema:
            $ref: '#/definitions/responseWithMessage'
        400:
          description: "Invalid data"
          schema:
            $ref: '#/definitions/responseWithMessage'
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      x-swagger-router-controller: "Attendance"
definitions:
  '404':
    type: object
    additionalProperties: false
    properties:
      error:
        type: string
    required:
      - error
  responseWithMessage:
    type: object
    additionalProperties: false
    properties:
      message:
        type: string
    required:
      - message
...
  beaconItem:
    type: object
    additionalProperties: false
    properties:
      id:
        type: integer
        format: int64
      mac_address:
        description: The MAC address of the Bluetooth point. Address with
            colon (:) or hyphen (-) separators, upper or lower case hex digits.
        type: string
        pattern: '^(([a-zA-Z0-9]{2}:){5}[a-zA-Z0-9]{2})|(([a-zA-Z0-9]{2}-){5}[a-zA-Z0-9]{2})$'
      active_status:
        type: boolean
    required:
      - id
      - mac_address
      - active_status
  beaconsList:
    type: array
    items:
      $ref: '#/definitions/beaconItem'
...
  beaconIds:
    type: array
    items:
      type: object
      properties:
        beaconId:
          type: integer
          format: int32
      required:
      - beaconId
      additionalProperties: false
    example:
      [
        {
          beaconId: 1
        },
        {
          beaconId: 2
        }
      ]
  roomIdBeaconId:
    type: object
    properties:
      roomId:
        type: integer
        format: int32
      beaconId:
        type: integer
        format: int32
    required:
    - roomId
    - beaconId
    additionalProperties: false
    example: {
      roomId: 1,
      beaconId: 1
    }
  allAssociations:
    type: array
    items:
      type: object
      properties:
        beacon_id:
          type: integer
          format: int32
        mac_address:
          type: string
        room_id:
          type: integer
          format: int32
        room_name:
          type: string
      required:
      - beacon_id
      - mac_address
      - room_id
      - room_name
    example: [
      {
        "beacon_id": 1,
        "mac_address": '7C:D1:C3:19:BC:EE',
        "room_id": 1,
        "room_name": "TC-110"
      },
      {
        "beacon_id": 2,
        "mac_address": '7C:D1:C3:19:BC:EA',
        "room_id": 1,
        "room_name": "TC-110"
      }
    ]
...
  arrayBeaconsApp:
    type: array
    items:
      type: string
    example: [
      '7C:D1:C3:19:BC:EE',
      '7C:D1:C3:19:BC:EF'
    ]
  filteredBeacons:
    type: array
    items:
      type: object
      properties:
        course_name:
          type: string
        room_name:
          type: string
        event_id:
          type: integer
          format: int32
      required:
      - course_name
      - room_name
      - event_id
      additionalProperties: false
    example: [
      {
        room_name: 'TC-110',
        course_name: 'Data Structures and Algorithms',
        event_id: 10
      }
    ]
...
  takeAttendanceBody:
    type: object
    properties:
      eventId:
        type: integer
        format: int32
      studentId:
        type: integer
        format: int32
    required:
    - eventId
    - studentId
    example: {
      eventId: 1,
      studentId: 1
    }
...
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "X-API-KEY"
    in: "header"